+++ main/fastcgi.h
@@ -31,7 +31,7 @@
  */
 
 #define FCGI_HASH_FUNC(var, var_len) \
-	(UNEXPECTED(var_len < 3) ? var_len : \
+	(UNEXPECTED(var_len < 3) ? (unsigned int)var_len : \
 		(((unsigned int)var[3]) << 2) + \
 		(((unsigned int)var[var_len-2]) << 4) + \
 		(((unsigned int)var[var_len-1]) << 2) + \
